+++ Coclustering/Ticker.h
@@ -10,7 +10,7 @@ public:
 	void start() {
 		mTStart = boost::posix_time::microsec_clock::local_time();
 	}
-	int stop() {
+	int64_t stop() {
 		mTStop = boost::posix_time::microsec_clock::local_time();
 		boost::posix_time::time_duration diff = mTStop - mTStart;
 		return diff.total_milliseconds();
